Implement time-based reminder with repeats
Finally. I think. Completely uncertain about how it really is.

Workmanager is downgraded to 0.2.0 (specific version) to avoid its iOS
problems (mentioned in previous commits) and adopt its Android
integration thing (idk what it does, but the setup is easier).

TimeReminderCenter is implemented to centralize all registration of
time-based reminders due to the limitation of the number of scheduled
notifications in iOS (64) and Android (50).

Now, during the launching of app and background fetches, reminder
registration is carried out. Time-based reminders with dates older than
the current date is updated to the first next matched date or is
disabled if the app cannot find the next one. First 64 or 50
notifications are then scheduled.

Registration of reminders slows down the app. Names editing in
ReminderScreen lags. Therefore, reminders are registered when the text
field is submitted instead of changed by characters.